refactored to use a factory that spits out response methods
