+++ index.js
@@ -3,9 +3,11 @@
 // Started from the example script in the node-slack-client project.
 
 var Slack = require('slack-client');
-var rhyme = require('rhyme-plus');
 var Q = require('q');
 var config = require('./config.js');
+var ResponseFactory = require('./response_factory.js');
+// make an instance to be re-used whenever a message comes in
+var responseFactory = ResponseFactory();
 
 // get the token from config
 var token = config.token;
@@ -36,97 +38,18 @@ var respondToText = function(channel, text) {
         // since some of our responses are asyncronous,
         // using a promise to simplify the code
         var deferred = Q.defer();
-        var message = '';
         
-        switch(possibleCommand) {
-            case 'rhyme':
-                // this lib uses a callback
-                rhyme(function (r) {
-
-                    // make this a little configurable
-                    var syllablePlaceholder = 'blah';
-
-                    // pop the last word to get a rhyming match (we'll have some fun with the rest)
-                    // not worried about changing the array of words since we're in a case statement
-                    var lastWord = messageWords.pop();
-
-                    // figure out how many syllables are left
-                    var remainingMessageSyllables = 0;
-                    // count is used in the loop below, count of syllables per word
-                    var count = 0;
-                    // loop over the words, and keep adding the count of syllables to the total
-                    for ( var j=0; j < messageWords.length; j++) {
-                        // get the number of syllables in each word
-                        count = r.syllables(messageWords[j]);
-                        if ( !!count ) {
-                            remainingMessageSyllables += count;                        
-                        }
-                    }
-
-                    // get an array of rhymes for the last word
-                    var rhymes = r.rhyme(lastWord);
-                    var numRhymes = rhymes.length;
-
-                    // gotta account for "orange"
-                    if ( !!numRhymes ) {
-                        // add a "blah" for each syllable (can be customized in plugin config)
-                        for ( var i=0; i < remainingMessageSyllables; i++) {
-                            message = syllablePlaceholder + ' ' + message;
-                        }
-                        // let's grab a random rhyming word from the array
-                        var randomSelection = Math.floor((Math.random() * (numRhymes - 1)));
-                        // and finally put a random rhyme after the blahs,
-                        // and low case it since for some reason the rhyme lib is LOUD...
-                        message = message + rhymes[randomSelection].toLowerCase();
-                        // resolve our promise with the message string
-                        deferred.resolve(message);
-                    } else {
-                        // if there is no rhyme for the last word, we'll punt
-                        message = 'No soup for you!';
-                        // resolve our promise with the message string
-                        deferred.resolve(message);
-                    }
-                });
-                break;
-            case 'missy':
-                // I put my thing down, flip it, and reverse it...
-                // ...ti esrever dna ,ti pilf ,nwod gniht ym tup I
-                message = messageWords.join(' ').split('').reverse().join('');
-                deferred.resolve(message);
-                break;
-            case 'echo':
-                // get the number of chars
-                // set a number of echos relative to that
-                // create n message strings
-                // replace more and more chars with empty spaces (using random position in the array)
-                // send it all back as a multiline message
-                var messageChars = messageWords.join(' ').split('');
-                var numChars = messageChars.length
-                var numEchoes = Math.floor(messageChars.length / 5);
-                var damagePerLoop = Math.floor(numChars / 4);
-                // first line in the response has no damage
-                var message = messageChars.join('');
-                var randomCharIdx = 0;
-                // repeat the message numEchoes times
-                for ( var k = 0; k <= numEchoes; k++ ) {
-                    for ( var l = 0; l <= damagePerLoop; l++ ) {
-                        // gete a random number
-                        randomCharIdx = Math.floor((Math.random() * (numChars - 1)));
-                        // replace a random char with a space
-                        messageChars[randomCharIdx] = '_';
-                    }
-                    var message = message + '\n' + messageChars.join('');
-                }
-                deferred.resolve(message);
-                break;
-            default:
-                // do nothing       
-        }
+        // use the response factory and pass the response method
+        // a promise and and array of words to work with
+        responseFactory.getResponse(possibleCommand)(deferred, messageWords);
 
         // when the promise is resolved, send a reply message
         deferred.promise.then(function (message) {
-            sendMessage(channel, message);
-            console.log('tinisi-chat-bot sending...');
+            // if message is anything other than an empty string, send a message
+            if ( message !== '' ) {
+                sendMessage(channel, message);
+                console.log('tinisi-chat-bot sending...');
+            }
         });
     }
 }
